requirements.txt
aiohttp>=3.9.0
pyyaml>=6.0
sqlalchemy>=2.0.0
aiosqlite>=0.19.0
slack-sdk>=3.23.0
apscheduler>=3.10.0
pydantic>=2.5.0
python-dotenv>=1.0.0
structlog>=23.2.0
